refactor: `lib/constant`에서 query key를 분리한다 (#134)
