+++ client/src/hooks/mutations/auth.ts
@@ -5,9 +5,8 @@ import type { LogInBody, LogInResult, SignUpBody, SignUpResult } from 'shared';
 import { useMutation } from '@tanstack/react-query';
 
 import { login, signUp } from '@/lib/api/auth';
-import { queryKey } from '@/lib/constants';
+import { queryKey } from '@/lib/queryKey';
 
-// TODO: Error 타입은 Api Error랑 error instanceof SyntaxError
 export const useSignUpMutation = (): UseMutationResult<
   SignUpResult,
   ApiError | SyntaxError,
