+++ client/src/lib/queryKey.ts
@@ -1,0 +1,8 @@
+const authQueryKey = {
+  signup: () => ['signup'] as const,
+  login: () => ['login'] as const,
+} as const;
+
+export const queryKey = {
+  auth: authQueryKey,
+};
